+++ lib/minitest-metadata.rb
@@ -1,6 +1,6 @@
 require 'minitest/metadata/version'
 require 'minitest/metadata'
 
-class MiniTest::Spec
-  include MiniTest::Metadata
+class Minitest::Spec
+  include Minitest::Metadata
 end