+++ lib/minitest/metadata.rb
@@ -1,4 +1,4 @@
-module MiniTest::Metadata
+module Minitest::Metadata
   module ClassMethods
     # Returns Hash metadata for class' test methods
     def metadata
@@ -31,7 +31,7 @@ module MiniTest::Metadata
         end
       end
 
-      # MiniTest 5 compatibility, https://github.com/wojtekmach/minitest-metadata/issues/7
+      # Minitest 5 compatibility, https://github.com/wojtekmach/minitest-metadata/issues/7
       alias_method :__name__, :name unless method_defined? :__name__
     end
     super